More build fixes.

+++ radio/src/gui/128x64/lcd.cpp
@@ -679,7 +679,7 @@ void drawSource(coord_t x, coord_t y, mixsrc_t idx, LcdFlags att)
   }
 #if defined(LUA_INPUTS)
   else if (idx <= MIXSRC_LAST_LUA) {
-    div_t qr = div(idx-MIXSRC_FIRST_LUA, MAX_SCRIPT_OUTPUTS);
+    div_t qr = div((uint16_t)(idx-MIXSRC_FIRST_LUA), MAX_SCRIPT_OUTPUTS);
 #if defined(LUA_MODEL_SCRIPTS)
     if (qr.quot < MAX_SCRIPTS && qr.rem < scriptInputsOutputs[qr.quot].outputsCount) {
       lcdDrawChar(x+2, y+1, '1'+qr.quot, TINSIZE);
